+++ ProjaraEJB/ejbModule/projara/model/dao/interfaces/ActionEventDaoLocal.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import projara.model.shop.ActionEvent;
 
-public interface ActionEventDaoLocal extends GenericDaoLocal<ActionEvent, Short> {
+public interface ActionEventDaoLocal extends GenericDaoLocal<ActionEvent, Integer> {
 
 	public List<ActionEvent> findActiveEvents();
 }
